requirements_gnn.txt
# Additional requirements for GNN training
torch-geometric>=2.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
requests>=2.31.0
lxml>=4.9.0
numpy>=1.24.0
